requirements.txt
fastapi
uvicorn
pydantic
pymongo[srv]
python-dotenv

